fix: 프로퍼티 주입 및 불필요 config 제거
