+++ src/main/kotlin/kr/hhplus/be/server/domain/concert/service/ConcertScheduleService.kt
@@ -7,11 +7,10 @@ import kr.hhplus.be.server.domain.concert.models.Seat
 import kr.hhplus.be.server.domain.concert.repositories.ConcertRepository
 import kr.hhplus.be.server.domain.concert.repositories.ConcertScheduleRepository
 import kr.hhplus.be.server.domain.common.ConcertBusinessRules
-import kr.hhplus.be.server.global.config.ConcertProperties
 import kr.hhplus.be.server.global.exception.ParameterValidationException
+import kr.hhplus.be.server.global.properties.ConcertProperties
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-
 
 @Service
 @Transactional
