+++ src/main/kotlin/kr/hhplus/be/server/global/config/BusinessRulesConfig.kt
@@ -2,10 +2,13 @@ package kr.hhplus.be.server.global.config
 
 import kr.hhplus.be.server.domain.common.ConcertBusinessRules
 import kr.hhplus.be.server.domain.common.ReservationBusinessRules
+import kr.hhplus.be.server.global.properties.ConcertProperties
+import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
+@EnableConfigurationProperties(ConcertProperties::class)
 class BusinessRulesConfig {
     
     @Bean
