+++ src/main/kotlin/kr/hhplus/be/server/global/config/JpaConfig.kt
@@ -1,18 +1,12 @@
 package kr.hhplus.be.server.global.config
 
-import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
-import org.springframework.orm.jpa.JpaTransactionManager
-import org.springframework.transaction.PlatformTransactionManager
+import org.springframework.transaction.annotation.EnableTransactionManagement
 
 @Configuration
 @EnableJpaAuditing
 @EnableJpaRepositories(basePackages = ["kr.hhplus.be.server"])
-class JpaConfig {
-    @Bean
-    fun transactionManager(): PlatformTransactionManager {
-        return JpaTransactionManager()
-    }
-}
+@EnableTransactionManagement
+class JpaConfig
